made basic refactor of the checkout pages
